use tentacles manager cli
