+++ tools/commands.py
@@ -18,23 +18,17 @@ import os
 import sys
 import asyncio
 import signal
-import aiohttp
 from threading import Thread
 from concurrent.futures import CancelledError
 
-from config import FORCE_ARG, ALL_ARG, UNINSTALL_ARG, UPDATE_ARG, INSTALL_ARG, HELP_ARG, DEFAULT_TENTACLES_URL
+from config import DEFAULT_TENTACLES_URL
 from octobot import get_bot, set_bot
 from octobot_commons.config_util import encrypt
 from octobot_tentacles_manager.api.configurator import get_tentacles_setup_config
-from octobot_tentacles_manager.api.creator import start_tentacle_creator
 from octobot_commons.logging.logging_util import get_logger
-from octobot_tentacles_manager.api.installer import install_all_tentacles, install_tentacles, \
-    USER_HELP as INSTALL_USER_HELP
 from octobot_tentacles_manager.api.loader import reload_tentacle_info
-from octobot_tentacles_manager.api.uninstaller import uninstall_all_tentacles, uninstall_tentacles, \
-    USER_HELP as UNINSTALL_USER_HELP
-from octobot_tentacles_manager.api.updater import update_all_tentacles, update_tentacles, \
-    USER_HELP as UPDATE_USER_HELP
+from octobot_tentacles_manager.cli import handle_tentacles_manager_command
+from tools.logger import init_logger
 
 COMMANDS_LOGGER_NAME = "Commands"
 
@@ -52,69 +46,9 @@ def data_collector(config, catch=True):
 #             raise e
 
 
-def _check_arg_and_cleanup(command_args, argument):
-    if argument in command_args:
-        command_args.pop(command_args.index(argument))
-        return True
-    return False
-
-
-async def _handle_package_manager_command(command_args, aiohttp_session=None):
-    aiohttp_session = aiohttp_session or aiohttp.ClientSession()
-    force = _check_arg_and_cleanup(command_args, FORCE_ARG)
-    all = _check_arg_and_cleanup(command_args, ALL_ARG)
-    install = _check_arg_and_cleanup(command_args, INSTALL_ARG)
-    update = _check_arg_and_cleanup(command_args, UPDATE_ARG)
-    uninstall = _check_arg_and_cleanup(command_args, UNINSTALL_ARG)
-
-    if not (all or command_args):
-        get_logger(COMMANDS_LOGGER_NAME).error("Please provide at least one tentacle name or add the 'all' parameter")
-    elif install:
-        if all:
-            await install_all_tentacles(DEFAULT_TENTACLES_URL, aiohttp_session=aiohttp_session)
-        else:
-            await install_tentacles(command_args, DEFAULT_TENTACLES_URL, aiohttp_session=aiohttp_session)
-    elif update:
-        if all:
-            await update_all_tentacles(DEFAULT_TENTACLES_URL, aiohttp_session=aiohttp_session)
-        else:
-            await update_tentacles(command_args, DEFAULT_TENTACLES_URL, aiohttp_session=aiohttp_session)
-    elif uninstall:
-        if all:
-            await uninstall_all_tentacles(DEFAULT_TENTACLES_URL, use_confirm_prompt=force)
-        else:
-            await uninstall_tentacles(command_args, DEFAULT_TENTACLES_URL, use_confirm_prompt=force)
-
-
-def _print_tentacle_manager_help():
-    help_message = f"""Welcome to OctoBot tentacles package manager.
-Available commands are:
-    - {INSTALL_ARG}: {INSTALL_USER_HELP}
-    - {UPDATE_ARG}: {UPDATE_USER_HELP}
-    - {UNINSTALL_ARG}: {UNINSTALL_USER_HELP}
-
-Protip: For each command, you can use the 'all' parameter to apply the command to all tentacles or apply the command to one 
-or many tentacles by naming them."""
-    print(help_message)
-
-
-def package_manager(command_args, catch=False, aiohttp_session=None):
-    try:
-        if _check_arg_and_cleanup(command_args, HELP_ARG):
-            _print_tentacle_manager_help()
-        else:
-            asyncio.run(_handle_package_manager_command(command_args, aiohttp_session))
-    except Exception as e:
-        if not catch:
-            raise e
-
-
-def tentacle_creator(config, commands, catch=False):
-    try:
-        start_tentacle_creator(config, commands)
-    except Exception as e:
-        if not catch:
-            raise e
+def call_tentacles_manager(command_args):
+    init_logger()
+    handle_tentacles_manager_command(command_args, tentacles_url=DEFAULT_TENTACLES_URL)
 
 
 def exchange_keys_encrypter(catch=False):
